requirements.txt
# ===== 核心依赖 =====
# Playwright - 浏览器自动化框架
playwright>=1.40.0

# LangGraph - AI Agent 编排框架
langgraph>=0.0.30

# LangChain - LangGraph 的依赖
langchain>=0.1.0
langchain-core>=0.1.0

# OpenAI - GPT-4o Vision API
openai>=1.12.0

# Python-dotenv - 环境变量管理
python-dotenv>=1.0.0

# ===== 开发依赖（可选） =====
# Python 类型检查
# mypy>=1.7.0

# 代码格式化
# black>=23.12.0

# 测试框架
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
